refactor navigation component
