+++ client/reader/discover/discover-navigation.scss
